Fix categorization issue
