+++ law.cfg
@@ -28,9 +28,9 @@ default_config: config_2022pre_limited
 default_dataset: tt_sl_powheg
 
 calibration_modules: columnflow.calibration.cms.{jets,met}, xyh.calibration.{default}
-selection_modules: columnflow.selection.{empty}, columnflow.selection.cms.{json_filter,met_filters}, xyh.selection.{default,categories}
-production_modules: columnflow.production.{categories,matching,normalization,processes}, columnflow.production.cms.{btag,electron,jet,matching,mc_weight,muon,pdf,pileup,scale,seeds}, xyh.production.{default,categories}
-categorization_modules: xyh.selection.categories, xyh.production.categories
+selection_modules: columnflow.selection.{empty}, columnflow.selection.cms.{json_filter,met_filters}, xyh.selection.{default}
+production_modules: columnflow.production.{categories,matching,normalization,processes}, columnflow.production.cms.{btag,electron,jet,matching,mc_weight,muon,pdf,pileup,scale,seeds}, xyh.production.{default}
+categorization_modules: xyh.categorization.example
 # TODO: Can be dropped if we apply SF in production?
 # TODO: Check, now it should be hist_production_module
 # weight_production_modules: columnflow.weight.{empty,all_weights}, xyh.weight.example
