+++ xyh/categorization/example.py
@@ -24,3 +24,32 @@ def cat_incl(self: Categorizer, events: ak.Array, **kwargs) -> tuple[ak.Array, a
 def cat_2j(self: Categorizer, events: ak.Array, **kwargs) -> tuple[ak.Array, ak.Array]:
     # two or more jets
     return events, ak.num(events.Jet.pt, axis=1) >= 2
+
+
+@categorizer(uses={"event"}, call_force=True)
+def catid_1e(self: Categorizer, events: ak.Array, **kwargs) -> tuple[ak.Array, ak.Array]:
+  mask = (ak.num(events.Electron, axis=-1) == 1) & (ak.num(events.Muon, axis=-1) == 0)
+  return events, mask
+
+
+@categorizer(uses={"event"}, call_force=True)
+def catid_1mu(self: Categorizer, events: ak.Array, **kwargs) -> tuple[ak.Array, ak.Array]:
+  mask = (ak.num(events.Electron, axis=-1) == 0) & (ak.num(events.Muon, axis=-1) == 1)
+  return events, mask
+
+
+@categorizer(uses={"Jet"}, call_force=True)
+def catid_6jets(self: Categorizer, events: ak.Array, **kwargs) -> tuple[ak.Array, ak.Array]:
+    mask = (ak.num(events.Jet, axis=-1) >= 6)
+    return events, mask
+
+
+@categorizer(uses={"Jet"}, call_force=True)
+def catid_5jets(self: Categorizer, events: ak.Array, **kwargs) -> tuple[ak.Array, ak.Array]:
+    mask = (ak.num(events.Jet, axis=-1) == 5)
+    return events, mask
+
+@categorizer(uses={"Jet"}, call_force=True)
+def catid_4jets(self: Categorizer, events: ak.Array, **kwargs) -> tuple[ak.Array, ak.Array]:
+    mask = (ak.num(events.Jet, axis=-1) == 4)
+    return events, mask
